requirements_dev.txt
locust
python-dotenv
pytest
dotenv-cli
pytest-asyncio